Move resolver to a dedicated package
