+++ requests/resolver/resolver.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package requests
+package resolver
 
 import (
 	"context"
