+++ requests/resolver/resolver_cache.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package requests
+package resolver
 
 import (
 	"sync"
